feat(ajustes): adiciona a funcionalidade de mudar o status ao completar 100% da leitura do livro na página editar-progresso; melhora o visual dos cards na pagina estatisticas; adiciona um confirm para a exclusão de livros na estante

+++ src/app/pages/estante/estante.page.ts
@@ -67,8 +67,14 @@ export class EstantePage implements OnInit {
   }
 
   removerLivro(event: Event, id: string) {
-    event.stopPropagation(); // Impede a navegação
-    this.livros = this.livros.filter(livro => livro.id !== id);
-    localStorage.setItem('estante', JSON.stringify(this.livros));
+    let conf = confirm("Deseja realmente excluir este livro?")
+    if(conf) {
+      event.stopPropagation(); // Impede a navegação
+      this.livros = this.livros.filter(livro => livro.id !== id);
+      localStorage.setItem('estante', JSON.stringify(this.livros));
+    } else {
+      event.stopPropagation();
+      return
+    }
   }
 }
